requirements.txt
mne>=1.6.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
prefect>=2.0.0
pytest>=7.0.0
snowflake-connector-python[pandas]>=3.5.0
ruff>=0.1.0